+++ storage/app/cipi/patch202112101.sh
@@ -1,0 +1,4 @@
+# NGINX RELOAD FIX
+sudo echo 'DefaultStartLimitIntervalSec=1s' >> /usr/lib/systemd/system/user@.service
+sudo echo 'DefaultStartLimitBurst=50' >> /usr/lib/systemd/system/user@.service
+sudo systemctl daemon-reload