+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Google Bot noindex, nofollow (for Panel)
 ### Fixed
+- Client Server Patch (build 202112101)
+- Site Domains and Paths in lowercase
+- SSL Nginx Restart Fix
 - PHP Version fix in site PDFs
 - Datatable Render Issue
 - Minor Template Fix